Enable eval with scope for web
