+++ packages/devtools_app/lib/src/shared/console/widgets/evaluate.dart
@@ -448,10 +448,6 @@ class ExpressionEvalFieldState extends State<ExpressionEvalField>
     if (assignment == null) return false;
     const kSuccess = true;
 
-    if (!evalService.isScopeSupported(emitWarningToConsole: true)) {
-      return kSuccess;
-    }
-
     final variable = serviceConnection.consoleService
         .itemAt(assignment.consoleItemIndex + 1);
     final value = variable?.value;
